Fase 3.6 (correções) e Fase 4 (scanner) completas — inclui extensão A01-A08, Meus Crafts e task 19 validada em jogo
Fecha as tasks que ainda valiam da Fase 3.6 "correções" (05, 06, 08, 09, 10, 13, 14,
15, 17) sobre a base já entregue pela Fase 4 (scanner: Market Flip, Refino, Craft,
Comida & Poções, Calculadora, e a extensão A01-A08 — Meus Crafts, marca do produto,
qualidade no craft, cenário individual por craft).

Destaques da Fase 3.6:
- Contrato HTTP 100% em inglês (query params/enums), tipo único de dinheiro em toda
  a API, índice trigram de busca de item comprovado no plano de consulta real.
- Serving de produção do frontend (Dockerfile multi-stage + nginx, mesma origem
  sem CORS) e "Abrir Calculadora" no systray do client Go.
- Baseline Git verificado de verdade pelo gate (scripts/verify_repository.py),
  formatação do client, decisão sobre a reconciliação 20.4, higiene de imagem
  Docker e do repositório (dockerignore, W9/W10/W11 do achados da Fase 3.5).

Task 19 (ensaio integrado em jogo, o gate final da Fase 3) concluída: validação
real contra as imagens de produção com Albion Online aberto, coleta do client
real, Meus Crafts e Calculadora conferidos ponta a ponta. Corrigido no processo:
HEALTHCHECK do frontend usava `localhost`, que resolve para `::1` dentro do
container Docker e falhava mesmo com o nginx respondendo normalmente — trocado
para `127.0.0.1`.

Docs atualizados para refletir o fechamento: Fase 3 sobe de 18/19 para 19/19,
checklist da Fase 3.6 marcado, CLAUDE.md/AGENTS.md/README.md com o estado real.

+++ backend/src/craft/schemas.py
@@ -90,7 +90,7 @@ class IngredientSimulationOut(BaseModel):
 
 
 class RecipeSimulationOut(BaseModel):
-    silver_cost_per_execution: int
+    silver_cost_per_execution: Decimal
     crafting_focus_per_execution: int
     amount_crafted: int
